Simplify GDAL installation.

+++ src/main/scala/com/databricks/labs/mosaic/gdal/MosaicGDAL.scala
@@ -2,33 +2,44 @@ package com.databricks.labs.mosaic.gdal
 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.SparkException
 import org.gdal.gdal.gdal
 
-import java.io.{BufferedInputStream, IOException}
+import java.io.BufferedInputStream
 import java.nio.file.{Files, Paths}
-import scala.io.{BufferedSource, Source}
 import scala.language.postfixOps
 import scala.sys.process._
 import scala.util.Try
 
 object MosaicGDAL extends Logging {
 
-    val GDAL_ENABLED = "spark.mosaic.gdal.native.enabled"
+    private val GDAL_ENABLED = "spark.mosaic.gdal.native.enabled"
+
+    private val mosaicGDALPath = Files.createTempDirectory("mosaic-gdal")
+    private val mosaicGDALAbsolutePath = mosaicGDALPath.toAbsolutePath.toString
+
+    if (!Files.exists(mosaicGDALPath)) Files.createDirectories(mosaicGDALPath)
 
     private var isEnabled = false
 
     private def wasEnabled(spark: SparkSession): Boolean = spark.conf.get(GDAL_ENABLED, "false").toBoolean
 
+    private def readResourceFile(name: String): Array[Byte] = {
+        val bis = new BufferedInputStream(getClass.getResourceAsStream(name))
+        try Stream.continually(bis.read()).takeWhile(-1 !=).map(_.toByte).toArray
+        finally bis.close()
+    }
+
+    def prepareEnvironment(scriptPath: String = "/FileStore/geospatial/mosaic/gdal/init_scripts/"): Unit = {
+        copySharedObjects()
+        copyInstallScript(scriptPath)
+    }
+
     def enableGDAL(spark: SparkSession): Unit = {
         if (!wasEnabled(spark) && !isEnabled) {
             isEnabled = true
-            installGDAL(spark)
-            copySharedObjects()
             loadSharedObjects()
             gdal.AllRegister()
             spark.conf.set(GDAL_ENABLED, "true")
-            spark.sparkContext.addSparkListener(new GDALListener)
         }
     }
 
@@ -53,25 +64,20 @@ object MosaicGDAL extends Logging {
         }
     }
 
-    def copySharedObjects(): Unit = {
-        def readSOBytes(name: String): Array[Byte] = {
-            val bis = new BufferedInputStream(getClass.getResourceAsStream(name))
-            try Stream.continually(bis.read()).takeWhile(-1 !=).map(_.toByte).toArray
-            finally bis.close()
-        }
+    private def copyInstallScript(path: String = "/FileStore/geospatial/mosaic/gdal/init_scripts/"): Unit = {
+        val scriptBytes = readResourceFile("/script/install-gdal-databricks.sh")
+        Files.write(Paths.get(s"$mosaicGDALAbsolutePath/mosaic-gdal.sh"), scriptBytes)
+        s"sudo cp $mosaicGDALAbsolutePath/mosaic-gdal.sh $path/mosaic-gdal.sh".!!
+    }
 
-        val so = readSOBytes("/gdal/ubuntu/libgdalalljni.so")
-        val so30 = readSOBytes("/gdal/ubuntu/libgdalalljni.so.30")
+    private def copySharedObjects(): Unit = {
+        val so = readResourceFile("/gdal/ubuntu/libgdalalljni.so")
+        val so30 = readResourceFile("/gdal/ubuntu/libgdalalljni.so.30")
 
-        val mosaicGDALPath = Files.createTempDirectory("mosaic-gdal")
-        val mosaicGDALAbsolutePath = mosaicGDALPath.toAbsolutePath.toString
         val usrGDALPath = Paths.get("/usr/lib/jni/")
         val libgdalSOPath = Paths.get("/usr/lib/libgdal.so")
-        if (!Files.exists(mosaicGDALPath)) Files.createDirectories(mosaicGDALPath)
         if (!Files.exists(usrGDALPath)) Files.createDirectories(usrGDALPath)
-        if (!Files.exists(libgdalSOPath)) {
-            "sudo cp /usr/lib/libgdal.so.30 /usr/lib/libgdal.so".!!
-        }
+        if (!Files.exists(libgdalSOPath)) "sudo cp /usr/lib/libgdal.so.30 /usr/lib/libgdal.so".!!
         Files.write(Paths.get(s"$mosaicGDALAbsolutePath/libgdalalljni.so"), so)
         Files.write(Paths.get(s"$mosaicGDALAbsolutePath/libgdalalljni.so.30"), so30)
 
@@ -79,46 +85,12 @@ object MosaicGDAL extends Logging {
         s"sudo cp $mosaicGDALAbsolutePath/libgdalalljni.so.30 /usr/lib/jni/libgdalalljni.so.30".!!
     }
 
-    def loadSharedObjects(): Unit = {
+    private def loadSharedObjects(): Unit = {
         System.load("/usr/lib/libgdal.so.30")
         System.load("/usr/lib/jni/libgdalalljni.so.30")
         System.load("/usr/lib/libgdal.so.30.0.3")
         System.load("/usr/lib/ogdi/libgdal.so")
         System.load("/usr/lib/libgdal.so")
-    }
-
-    def installGDAL(spark: SparkSession): Unit = installGDAL(Some(spark))
-
-    def installGDAL(spark: Option[SparkSession]): Unit = {
-        val sc = spark.map(_.sparkContext)
-        val numExecutors = sc.map(_.getExecutorMemoryStatus.size - 1)
-        val script = getScript
-        for (cmd <- script.getLines.toList) {
-            try {
-                if (!cmd.startsWith("#") || cmd.nonEmpty) cmd.!!
-                sc.map { sparkContext =>
-                    if (!sparkContext.isLocal) {
-                        sparkContext.parallelize(1 to numExecutors.get).pipe(cmd).collect
-                    }
-                }
-            } catch {
-                case e: IOException           => logError(e.getMessage)
-                case e: IllegalStateException => logError(e.getMessage)
-                case e: SparkException        => logError(e.getMessage)
-                case e: Throwable             => logError(e.getMessage)
-            } finally {
-                script.close
-            }
-        }
-    }
-
-    def getScript: BufferedSource = {
-        val scriptPath = System.getProperty("os.name").toLowerCase() match {
-            case o: String if o.contains("nux") => "/scripts/install-gdal-databricks.sh"
-            case _ => throw new UnsupportedOperationException("This method only supports Ubuntu Linux with `apt`.")
-        }
-        val script = Source.fromInputStream(getClass.getResourceAsStream(scriptPath))
-        script
     }
 
 }
